+++ scripts/setup.sh
@@ -8,17 +8,22 @@
 #     ./scripts/setup.sh [ARGS]
 #
 # Args:
-#     1 - Path for the parent directory of the virtual environment called "venv"
+#     1 - Path for the parent directory of the virtual environment called "venv" (DEFAULT = $HOME)
 #
 #
 #
 
+# Ensure running from the repo
+repoDir=$(git rev-parse --show-toplevel)
+if [ $? -ne 0 ]; then echo "ERROR: Run this command from the homer git repo"; exit 1; fi
+cd ${repoDir}/
+
+# Arg for venv location
 defaultVenvLocation="${HOME}"
-workingDir=$(pwd)
 userSpecifiedVenvLocation="${1}"
 
 # Ensure the script is running from the pycons3rt3 directory
-if [ ! -f ${workingDir}/pycons3rt3/VERSION.txt ]; then
+if [ ! -f ${repoDir}/pycons3rt3/VERSION.txt ]; then
     echo "ERROR: Please run this script from the top-level driectory of the pycons3rt3 git repo"
     exit 1
 fi
@@ -31,10 +36,6 @@ if [ $? -ne 0 ]; then
     echo "python3 not detected, please install python3 to continue"
     exit 1
 fi
-
-echo "Installing virtualenv..."
-python3 -m pip install virtualenv
-if [ $? -ne 0 ]; then echo "Problem installing the virtualenv package from pip"; exit 2; fi
 
 if [ -z "${userSpecifiedVenvLocation}" ]; then
     read -p "Type path to parent directory for the virtual environment venv directory [default: ${HOME}]" userSpecifiedVenvLocation
@@ -58,21 +59,13 @@ fi
 # Create the virtual environment
 cd ${parentDir}/
 echo "Creating venv..."
-python3 -m virtualenv venv
+python3 -m venv venv
 if [ $? -ne 0 ]; then echo "Problem creating venv in directory: ${parentDir}"; exit 4; fi
 
 # Activate the virtual environment
 echo "Activating the virtual environment..."
-. ./venv/bin/activate
+. ${parentDir}/venv/bin/activate
 if [ $? -ne 0 ]; then echo "Problem activating the virtual environment in: ${parentDir}/venv"; exit 5; fi
-
-# Print the python3 in use
-echo "Using python3: "
-which python3
-
-# Print the version
-echo "Using python3 version: "
-python3 --version
 
 # upgrade pip in the venv
 echo "Upgrading pip in the virtual environment..."
@@ -80,14 +73,14 @@ python3 -m pip install pip --upgrade
 if [ $? -ne 0 ]; then echo "Problem upgrading pip in the virtual environment: ${parentDir}/venv"; exit 6; fi
 
 # Install prereqs
-cd ${workingDir}/
+cd ${repoDir}/
 echo "Installing prerequisites..."
-python3 -m pip install -r ./cfg/requirements.txt
-if [ $? -ne 0 ]; then echo "Problem installing prerequisite packages from cfg/requirements.txt"; exit 7; fi
+python3 -m pip install -r ./requirements.txt
+if [ $? -ne 0 ]; then echo "Problem installing prerequisite packages from requirements.txt"; exit 7; fi
 
 # Install pycons3rt3
 echo "Installing pycons3rt3..."
-python3 setup.py install
+python3 -m pip install .
 if [ $? -ne 0 ]; then echo "Problem installing pycons3rt3"; exit 8; fi
 
 echo "Completed installation of pycons3rt3 into virtual environment: ${parentDir}/venv"